write json via cfg reports
