+++ pacman/utilities/file_format_schemas/machine.json
